Refactors docker login functionality out of docker_cache script and re-uses it for CD
